modules/persons-service/requirements.txt
grpcio_tools==1.48.1
grpcio==1.48.1
SQLAlchemy==1.3.19
GeoAlchemy2==0.8.4
psycopg2-binary==2.9.3
